backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiohttp>=3.11.0
pydantic>=2.10.0
python-dotenv>=1.0.0
